Fixes for multiple vagrant provisions

+++ scripts/puppet_install.sh
@@ -2,14 +2,28 @@
 eval `cat /etc/lsb-release`
 sudo wget -qO - https://apt.puppetlabs.com/pubkey.gpg | sudo apt-key add -
 echo -n "* added puppetlabs apt key"
-sudo echo "deb https://apt.puppetlabs.com/ $(sudo lsb_release -c -s) PC1" | sudo tee -a /etc/apt/sources.list.d/puppet.list
-echo -n "* added puppetlabs apt repo"
+if [[ ! -e "/etc/apt/sources.list.d/puppet.list" ]];
+  then
+    sudo echo "deb https://apt.puppetlabs.com/ $(sudo lsb_release -c -s) PC1" | sudo tee -a /etc/apt/sources.list.d/puppet.list
+    echo -n "* added puppetlabs apt repo"
+else
+  if ! grep -q "deb https://apt.puppetlabs.com/ $(sudo lsb_release -c -s) PC1" /etc/apt/sources.list.d/puppet.list;
+   then
+       sudo echo "deb https://apt.puppetlabs.com/ $(sudo lsb_release -c -s) PC1" | sudo tee -a /etc/apt/sources.list.d/puppet.list
+       echo -n "* added puppetlabs apt repo"
+   else
+     echo -n "* repo already added... moving on"
+
+  fi
+fi
 echo -n "* Executing apt-get update"
 sudo apt-get update
 echo -n "* Attempting to install puppet"
 sudo apt-get install -y --force-yes puppet-agent=1.8.*
 echo " - Done"
-sudo ln -s /opt/puppetlabs/bin/puppet /usr/bin/puppet
+if [[ ! -e /usr/bin/puppet ]]; then
+  sudo ln -s /opt/puppetlabs/bin/puppet /usr/bin/puppet
+fi
 sudo /usr/bin/logger -t autobootstrap "installed puppet"
 echo "* installing gems"
 sudo /opt/puppetlabs/puppet/bin/gem install hiera-eyaml
